Moving tests to underscore folder Adding ConsoleLog tests #31
