+++ Tests/Tests.Strasnote.Logging.Serilog/Tests.Strasnote.Logging.Serilog.csproj
@@ -20,4 +20,8 @@
     <ProjectReference Include="..\..\Util\Strasnote.Util.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Extensions\" />
+  </ItemGroup>
+
 </Project>
